consolidated package structure.
