+++ ServiceApp/src/org/droidplanner/services/android/core/drone/companion/solo/artoo/ArtooLinkListener.java
@@ -1,6 +1,6 @@
-package org.droidplanner.services.android.drone.companion.solo.artoo;
+package org.droidplanner.services.android.core.drone.companion.solo.artoo;
 
-import org.droidplanner.services.android.drone.companion.solo.AbstractLinkManager;
+import org.droidplanner.services.android.core.drone.companion.solo.AbstractLinkManager;
 import com.o3dr.services.android.lib.drone.companion.solo.button.ButtonPacket;
 import com.o3dr.services.android.lib.drone.companion.solo.tlv.TLVPacket;
