+++ ServiceApp/src/org/droidplanner/services/android/utils/CommonApiUtils.java
@@ -74,7 +74,7 @@ import org.droidplanner.services.android.core.helpers.coordinates.Coord3D;
 import org.droidplanner.services.android.core.mission.survey.SplineSurveyImpl;
 import org.droidplanner.services.android.core.mission.survey.SurveyImpl;
 import org.droidplanner.services.android.core.survey.Footprint;
-import org.droidplanner.services.android.drone.companion.solo.SoloComp;
+import org.droidplanner.services.android.core.drone.companion.solo.SoloComp;
 import org.droidplanner.services.android.utils.file.IO.ParameterMetadataLoader;
 import org.xmlpull.v1.XmlPullParserException;
